% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crowdpen.R
\docType{data}
\name{crowdpen_api_url}
\alias{crowdpen_api_url}
\title{Crowdpen API base URL}
\format{\preformatted{ chr "https://api.crowdpen.com/api/rest/v1/results"
}}
\usage{
crowdpen_api_url
}
\arguments{
\item{crowdpen_api_url}{Crowdpen API base url, default is 'https://api.crowdpen.com/api/rest/v1/results'}
}
\description{
Crowdpen API base URL
}
\keyword{datasets}

