% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crowdpen.R
\docType{data}
\name{crowdpen_email}
\alias{crowdpen_email}
\title{Crowdpen API username}
\format{\preformatted{ chr "michiel@nosuchtype.com"
}}
\usage{
crowdpen_email
}
\arguments{
\item{crowdpen_email}{Crowdpen API username - required, no default}
}
\description{
Crowdpen API username
}
\keyword{datasets}

