% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crowdpen.R
\docType{data}
\name{crowdpen_token}
\alias{crowdpen_token}
\title{Crowdpen API base URL}
\format{\preformatted{ chr "7DNcm-QLwHhirkssQFCS"
}}
\usage{
crowdpen_token
}
\arguments{
\item{crowdpen_token}{Crowdpen API access token, required, no default}
}
\description{
Crowdpen API base URL
}
\keyword{datasets}

