% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.R
\name{fetch_cp_data}
\alias{fetch_cp_data}
\title{Fetch crowdpen data}
\usage{
fetch_cp_data(qid, sortDir = "asc", sortBy = "raised", page = 1,
  pageSize = 100)
}
\arguments{
\item{qid}{Searchquery ID, required.}
}
\description{
This function allows you to fetch data from the crowdpen.com api.
}
\examples{
cp_fetch(qid)
}
\keyword{crowdfunding}

