% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crowdpen.R
\name{fetch_results}
\alias{fetch_results}
\title{Fetch crowdpen data}
\usage{
fetch_results(qid, sortDir = "desc", sortBy = "raised", page = 1,
  pageSize = 100)
}
\arguments{
\item{qid}{Searchquery ID, required.}

\item{sortDir,}{sort direction, String, default='desc'}

\item{sortBy,}{sort direction, String, default='raised'}

\item{page,}{page number, Number, default='1'}

\item{pageSize,}{elements in page, Number, default='100'}
}
\description{
This function allows you to fetch data from the crowdpen.com api.
}
\examples{
fetch_results(qid)
}
\keyword{crowdfunding}

