% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crowdpen.R
\name{set_auth_email}
\alias{set_auth_email}
\title{Set crowdpen username/email address}
\usage{
set_auth_email(email)
}
\arguments{
\item{email}{email address, required}
}
\description{
This function allows you to set your crowdpen.com api email.
}
\examples{
set_auth_email()
}

