% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crowdpen.R
\name{set_auth_token}
\alias{set_auth_token}
\title{Set crowdpen auth token}
\usage{
set_auth_token(token)
}
\arguments{
\item{token}{token, required}
}
\description{
This function allows you to set your crowdpen.com api auth token.
}
\examples{
set_auth_token()
}

