% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crowdpen.R, R/test.R
\name{fetch_cp_data}
\alias{fetch_cp_data}
\title{Fetch crowdpen data}
\usage{
fetch_cp_data(qid, sortDir = "asc", sortBy = "raised", page = 1,
  pageSize = 100)

fetch_cp_data(qid, sortDir = "asc", sortBy = "raised", page = 1,
  pageSize = 100)
}
\arguments{
\item{qid}{Searchquery ID, required.}

\item{sortDir,}{sort direction, String, default='desc'}

\item{sortBy,}{sort direction, String, default='raised'}

\item{page,}{page number, Number, default='1'}

\item{pageSize,}{elements in page, Number, default='100'}

\item{qid}{Searchquery ID, required.}
}
\description{
This function allows you to fetch data from the crowdpen.com api.

This function allows you to fetch data from the crowdpen.com api.
}
\examples{
fetch_cp_data(qid)
cp_fetch(qid)
}
\keyword{crowdfunding}

